Move permission logic into dedicated module
