+++ packages/server/src/dataset/dataset.module.ts
@@ -4,10 +4,11 @@ import { DatasetService } from './dataset.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { Dataset, DatasetSchema } from './dataset.model';
 import { DatasetPipe } from './pipes/dataset.pipe';
-import { AuthModule } from '../auth/auth.module';
+import { PermissionModule } from '../permission/permission.module';
+import { JwtModule } from '../jwt/jwt.module';
 
 @Module({
-  imports: [MongooseModule.forFeature([{ name: Dataset.name, schema: DatasetSchema }]), AuthModule],
+  imports: [MongooseModule.forFeature([{ name: Dataset.name, schema: DatasetSchema }]), PermissionModule, JwtModule],
   providers: [DatasetResolver, DatasetService, DatasetPipe],
   exports: [DatasetService, DatasetPipe]
 })
