+++ packages/server/src/entry/resolvers/entry.resolver.ts
@@ -5,8 +5,8 @@ import { EntryService } from '../services/entry.service';
 import { DatasetPipe } from '../../dataset/pipes/dataset.pipe';
 import { UseGuards, Inject, UnauthorizedException } from '@nestjs/common';
 import { JwtAuthGuard } from '../../jwt/jwt.guard';
-import { DatasetPermissions } from '../../auth/permissions/dataset';
-import { CASBIN_PROVIDER } from '../../auth/casbin.provider';
+import { DatasetPermissions } from '../../permission/permissions/dataset';
+import { CASBIN_PROVIDER } from '../../permission/casbin.provider';
 import * as casbin from 'casbin';
 import { TokenPayload } from '../../jwt/token.dto';
 import { TokenContext } from '../../jwt/token.context';
