+++ packages/server/src/entry/resolvers/upload-session.resolver.ts
@@ -7,8 +7,8 @@ import { UploadSessionPipe } from '../pipes/upload-session.pipe';
 import { DatasetPipe } from '../../dataset/pipes/dataset.pipe';
 import { UploadResult } from '../dtos/upload-result.dto';
 import { JwtAuthGuard } from '../../jwt/jwt.guard';
-import { DatasetPermissions } from '../../auth/permissions/dataset';
-import { CASBIN_PROVIDER } from '../../auth/casbin.provider';
+import { DatasetPermissions } from '../../permission/permissions/dataset';
+import { CASBIN_PROVIDER } from '../../permission/casbin.provider';
 import * as casbin from 'casbin';
 import { TokenContext } from '../../jwt/token.context';
 import { TokenPayload } from '../../jwt/token.dto';
