+++ packages/server/src/permission/permission.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { casbinProvider } from './casbin.provider';
+
+@Module({
+  providers: [
+    casbinProvider
+  ],
+  exports: [casbinProvider]
+})
+export class PermissionModule {}
