+++ packages/server/src/project/project.module.ts
@@ -6,7 +6,8 @@ import { Project, ProjectSchema } from './project.model';
 import { ProjectPipe } from './pipes/project.pipe';
 import { MongooseMiddlewareService } from 'src/shared/service/mongoose-callback.service';
 import { SharedModule } from 'src/shared/shared.module';
-import { AuthModule } from '../auth/auth.module';
+import { JwtModule } from '../jwt/jwt.module';
+import { PermissionModule } from '../permission/permission.module';
 
 @Module({
   imports: [
@@ -27,7 +28,8 @@ import { AuthModule } from '../auth/auth.module';
         inject: [MongooseMiddlewareService]
       }
     ]),
-    AuthModule
+    JwtModule,
+    PermissionModule
   ],
   providers: [ProjectResolver, ProjectService, ProjectPipe],
   exports: [ProjectPipe, ProjectService]
